requirements.txt
altgraph==0.17.3
anyio==3.6.2
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
darkdetect==0.8.0
fastapi==0.95.1
fastdeploy-python==1.0.6
fastdeploy-tools==0.0.5
fonttools==4.39.3
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pefile==2023.2.7
Pillow==9.5.0
pydantic==1.10.7
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pyparsing==3.0.9
pyperclip==1.8.2
PyQt-Fluent-Widgets==0.8.4
PyQt5==5.15.9
PyQt5-Frameless-Window==0.2.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dateutil==2.8.2
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.29.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.16.0
zipp==3.15.0
